Fix the jest unit tests.  The logic was not quite correct for the setting of defaults in parserParams.
Also it is more proper to compare _params[key] to `undefined` rather
than `null` given that the type of the value of _params[key] is
technically `genric | Dictionary<generic> | undefined`.  `null` works
because the comparisons are not strict.  That is something that we
should probably investigate.  The dubious test in courses.spec.ts
allowing the passage of a `null` value indicates that we are really
allowing `null`.  If we are expecting that then the type should include
`null`.